chore: api 호출 관련 파일 위치 수정
